Move everything SWT-related from `Displayable` to `Screen`
